Move from async to sync file reading
Due to the async fs.writeFile, lobbies file was being overwritten with the same data multiple times.
This was causing the lobbies file to be corrupted and the server to crash.

I'm not sure nor have needed knowledge to know will changing to sync will fix the problem.
On paper, it should fix it. But I'm not sure if it will cause other problems.

If it does, I've probably will be forced to use a database or create a queue system for the file writing.